+++ packages/core/src/features/grasshopper/index.ts
@@ -1,5 +1,0 @@
-export * from './compute';
-export * from './io';
-export * from './types';
-export * from './client';
-export * from './data-tree';