+++ packages/core/src/features/grasshopper/io/input/index.ts
@@ -1,9 +1,0 @@
-/**
- * Input processing and validation - convenience re-exports
- */
-
-// Input validators
-export * from './input-validators';
-
-// Input parsers
-export { processInput, processInputs } from './input-processors';