+++ packages/core/src/features/visualization/threejs/index.ts
@@ -1,3 +1,0 @@
-export * from './three-helpers.js';
-export * from './three-initializer.js';
-export * as Materials from './three-materials.js';